Manage integration test IAM policy using Python instead of terraform
